requirements.txt
boto3>=1.28.0
ipywidgets>=8.0.0
Pillow>=9.0.0
python-docx>=0.8.11
pandas>=1.5.0
openpyxl>=3.0.0
numpy>=1.20.0
